More checks before deleting a nic.
* We shouldn't be able to delete a nic that's connected to a network
otherwise we get a 500 due to the attachments table having to refer to a non
existing nic.

+++ hil/api.py
@@ -355,11 +355,15 @@ def node_delete_nic(node, nic):
     """Delete nic with given name from it's node.
 
     If the node or nic does not exist, a NotFoundError will be raised.
-    If there's a pending networking action then a BlockedError will be raised.
+    If there's a pending networking action or if the nic is connected to a
+    network then a BlockedError will be raised.
     """
+
     get_auth_backend().require_admin()
     nic = get_child_or_404(get_or_404(model.Node, node), model.Nic, nic)
     check_pending_action(nic)
+    if nic.attachments:
+        raise errors.BlockedError("Nic is connected to a network")
     db.session.delete(nic)
     db.session.commit()
